\name{Division}
\alias{Division test}
\title{Division between two items}
\description{
}
\usage{
Division(x)
}
\arguments{
  \item{x}{
  a - numeric value
  b - numeric value
}
}
\details{
}
\value{
Give a numeric value 
}
\references{
}
\author{
Victoria Galvan Delgadillo
}
\note{
Clase de metaheuristicas
}


\seealso{
}
\examples{
 a <- 5
 b <- 2
{
  }
}

