\name{suma}
\alias{suma}
\title{Suma de Dos Numeros}
\description{La funcion suma dos numeros}
\usage{
suma(a, b)
}
\arguments{
  \item{a}{
    Primer número a sumar
    }
  \item{b}{
    Segundo número a sumar
    }
}
\description{
La funcion suma dos números
}
\value{
Regresa un valor numerico con el resultado de la suma del primer número y segundo número
}
\author{
Cristina
}
\examples{
suma (8,8)
}
