\name{division}
\alias{division}
\title{division de dos variables}

\usage{
division(x)
}

\arguments{
\item{a, b}{
  Son las variables que guardaran el numerador y denominador de la division a evaluar.
  }
}

\value{
Retorna un valor numerico, que es el resultado de la funcion division \code{x}.
}

\description{
La funcion division es una operacion aritmetica basica en repartir una cantidad total en partes iguales, obteniendo como resultado un cociente.
}

\examples{
# Division de 10 entre 5
division(10, 5)
}

\references{
{Hansen, N., Müller, S. D., & Koumoutsakos, P. (2003). \emph{Reducing the time complexity of the derandomized evolution strategy with covariance matrix adaptation (CMA-ES)}. Evolutionary Computation, 11(1), 1-18.}
}
